Отчёт обновлён

+++ Program/Mass/Source.cpp
@@ -4,7 +4,7 @@
 #include <iomanip>
 
 
-void inputdata(float arr1[], float arr2[], double n)
+void inputdata(float arr1[], float arr2[], int n)
 {
 	std::ifstream inputFile("InputData.csv");
 	float firstCol = 0;
@@ -15,14 +15,14 @@ void inputdata(float arr1[], float arr2[], double n)
 		arr2[i] = secondCol;
 	}
 }
-void mass(float arr1[], float arr2[], double n)
+void mass(float arr1[], float arr2[], int n)
 {
 	for (int i = 0; i < n; i++) {
 		arr1[i] = arr1[i] * (4 * 3.14 * pow(arr2[i] / 2, 3)) / 3;
 	}
 }
 
-void outputdata(float arr1[], float arr2[], double n)
+void outputdata(float arr1[], float arr2[], int n)
 {
 	mass(arr1, arr2, n);
 	std::ofstream outputFile("OutputData.csv");
@@ -31,7 +31,7 @@ void outputdata(float arr1[], float arr2[], double n)
 
 }
 int main() {
-	double n = 6;
+	int n = 6;
 	float data1[6];
 	float data2[6];
 	inputdata(data1, data2, n);
